ethdb/pebble: drop redundant int64 conversion in data-size gauge
pebble v2's LevelMetrics.TablesSize is already int64, so the int64()
conversion in the actual-data-size gauge is a no-op that golangci-lint's
unconvert linter rejects.

+++ ethdb/pebble/pebble.go
@@ -1052,7 +1052,7 @@ func (d *Database) updateCalculatedAmplifications(stats *pebble.Metrics) {
 	// level.TablesSize is the CURRENT live size, which already accounts for deleted files
 	var actualDataSize int64
 	for _, level := range stats.Levels {
-		actualDataSize += int64(level.TablesSize)
+		actualDataSize += level.TablesSize
 	}
 
 	d.actualDataSizeGauge.Update(actualDataSize)
